pyproject.toml
[project]
name = "ladder-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13.2,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "trio (>=0.29.0,<0.30.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "selenium (>=4.30.0,<5.0.0)",
    "django-ninja (>=1.4.0,<2.0.0)",
    "channels[daphne] (>=4.2.2,<5.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "uvloop (>=0.21.0,<0.22.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pytube (>=15.0.0,<16.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.0,<3.0.0)",
    "db-sqlite3 (>=0.0.1,<0.0.2)",
    "pysqlite3 (>=0.5.4,<0.6.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
packages = [{include = "ladder_project", from = "src"}]


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.11.8"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'python manage.py runserver'
run_worker = "celery -A ladder_server worker -l info"
run_beat = "celery -A ladder_server beat"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
